Setting up Nodemailer

+++ controllers/api/v1/posts_api.js
@@ -27,19 +27,22 @@ module.exports.destroy = async function(req, res)
     try
     {
         let post = await Post.findById(req.params.id);
-        // if(post.user == req.user.id)
-        // {
+        if(post.user == req.user.id)
+        {
             post.remove();
 
             await Comment.deleteMany({post: req.params.id});
             return res.json(200, {
                 message: 'Post and associated comments deleted successfully'
             });
-        // }
-        // else
-        // {
-        //     return res.redirect('back');
-        // }
+        }
+        else
+        {
+            return res.json(401, 
+            {
+                message: 'You cannot delete this post!'
+            });
+        }
     }
     catch(err)
     {
